Create `completions.rs`
